# contracts/openapi/core.v1.yaml
openapi: 3.1.0
info:
  title: ThomasAgent Core API
  version: 1.0.0
  description: >-
    All responses use the shared response format {data, meta} or {error}. Error codes come from contracts/errors.yaml.
    JSON request bodies require application/json (charset parameters are supported). Missing, malformed or other media types
    return UNSUPPORTED_MEDIA_TYPE (415). Unknown fields and malformed JSON return VALIDATION_FAILED (400).
    Requiring JSON and rejecting unknown fields are breaking changes for callers that previously relied on permissive decoding.
servers:
  - url: "{baseUrl}"
    variables:
      baseUrl: { default: http://localhost:8080 }
tags:
  - name: Health
  - name: Auth
  - name: Providers
  - name: Models
  - name: Model roles
  - name: Internal
security:
  - sessionCookie: []
    csrfHeader: []
paths:
  /healthz:
    get:
      tags: [Health]
      operationId: healthz
      security: []
      responses:
        "200": { $ref: "#/components/responses/HealthStatus" }
  /readyz:
    get:
      tags: [Health]
      operationId: readyz
      security: []
      responses:
        "200": { $ref: "#/components/responses/HealthStatus" }
        "503": { $ref: "#/components/responses/Error" }
  /api/v1/auth/login:
    post:
      tags: [Auth]
      operationId: login
      description: >-
        Login attempts share email/IP, email-global and IP-only budgets before password verification.
        Exhausting any budget returns RATE_LIMITED with Retry-After in seconds. The email-global
        budget can block a correct password for the remaining window; it has no password exemption.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
            example: { email: "{{ownerEmail}}", password: "{{ownerPassword}}" }
      responses:
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }
        "200":
          description: Signed in; sets thomas_session and thomas_csrf cookies.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LoginSuccessResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429":
          description: Login budget exhausted; wait the Retry-After duration before retrying.
          headers:
            Retry-After:
              description: Remaining blocked-window duration in seconds.
              schema: { type: integer, minimum: 1 }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "400": { $ref: "#/components/responses/Error" }
  /api/v1/auth/me:
    get:
      tags: [Auth]
      operationId: me
      description: Authenticates the stored session using database time for its expiry decision.
      security: [{ sessionCookie: [] }]
      responses:
        "200":
          description: Current user.
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/SuccessResponse" }
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          { user: { $ref: "#/components/schemas/User" } }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      operationId: logout
      responses:
        "200": { $ref: "#/components/responses/LogoutResult" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/providers:
    get:
      tags: [Providers]
      operationId: listProviders
      security: [{ sessionCookie: [] }]
      responses:
        "200": { $ref: "#/components/responses/ProviderList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [Providers]
      operationId: createProvider
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProviderInput" }
            example:
              {
                kind: openrouter,
                name: "OpenRouter {{$randomInt}}",
                baseUrl: "http://127.0.0.1:1",
              }
      responses:
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }
        "201": { $ref: "#/components/responses/Provider" }
        "400": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/providers/{providerId}:
    parameters:
      - {
          name: providerId,
          in: path,
          required: true,
          schema: { type: string, format: uuid },
          example: "{{providerId}}",
        }
    patch:
      tags: [Providers]
      operationId: updateProvider
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProviderInput" }
            example: { enabled: true }
      responses:
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }
        "200": { $ref: "#/components/responses/Provider" }
        "404": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "400": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
    delete:
      tags: [Providers]
      operationId: deleteProvider
      responses:
        "200": { $ref: "#/components/responses/DeleteResult" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "400": { $ref: "#/components/responses/Error" }
  /api/v1/providers/{providerId}/test:
    parameters:
      - {
          name: providerId,
          in: path,
          required: true,
          schema: { type: string, format: uuid },
          example: "{{providerId}}",
        }
    post:
      tags: [Providers]
      operationId: testProvider
      responses:
        "200": { $ref: "#/components/responses/ProviderTestResult" }
        "400": { $ref: "#/components/responses/Error" }
        "503": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/Error" }
  /api/v1/providers/{providerId}/sync:
    parameters:
      - {
          name: providerId,
          in: path,
          required: true,
          schema: { type: string, format: uuid },
          example: "{{providerId}}",
        }
    post:
      tags: [Providers]
      operationId: syncProvider
      responses:
        "200":
          description: Sync counts.
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/SuccessResponse" }
                  - {
                      type: object,
                      properties:
                        { data: { $ref: "#/components/schemas/SyncResult" } },
                    }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/Error" }
        "503": { $ref: "#/components/responses/Error" }
  /api/v1/models:
    get:
      tags: [Models]
      operationId: listModels
      parameters:
        - {
            name: providerId,
            in: query,
            schema: { type: string, format: uuid },
          }
        - {
            name: capability,
            in: query,
            schema: { $ref: "#/components/schemas/Capability" },
          }
      security: [{ sessionCookie: [] }]
      responses:
        "200": { $ref: "#/components/responses/ModelList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "400": { $ref: "#/components/responses/Error" }
    post:
      tags: [Models]
      operationId: createModel
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ModelInput" }
            example:
              {
                providerId: "{{providerId}}",
                modelRef: "openai/gpt-5-mini",
                displayName: "GPT-5 mini",
                capabilities: [chat],
              }
      responses:
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }
        "201": { $ref: "#/components/responses/Model" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /api/v1/models/{modelId}:
    parameters:
      - {
          name: modelId,
          in: path,
          required: true,
          schema: { type: string, format: uuid },
          example: "{{modelId}}",
        }
    delete:
      tags: [Models]
      operationId: deleteModel
      responses:
        "200": { $ref: "#/components/responses/DeleteResult" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "400": { $ref: "#/components/responses/Error" }
  /api/v1/model-roles:
    get:
      tags: [Model roles]
      operationId: listModelRoles
      security: [{ sessionCookie: [] }]
      responses:
        "200":
          description: Role assignments.
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/SuccessResponse" }
                  - {
                      type: object,
                      properties:
                        {
                          data:
                            {
                              type: array,
                              items:
                                { $ref: "#/components/schemas/RoleAssignment" },
                            },
                        },
                    }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/model-roles/{role}:
    parameters:
      - {
          name: role,
          in: path,
          required: true,
          schema: { $ref: "#/components/schemas/Role" },
          example: chat.fast,
        }
    put:
      tags: [Model roles]
      operationId: assignModelRole
      requestBody:
        required: true
        content:
          application/json:
            schema:
              {
                type: object,
                additionalProperties: false,
                required: [modelId],
                properties: { modelId: { type: string, format: uuid } },
              }
            example: { modelId: "{{modelId}}" }
      responses:
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }
        "200": { $ref: "#/components/responses/RoleAssignment" }
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /internal/models/resolve:
    get:
      tags: [Internal]
      operationId: resolveModel
      security: [{ serviceToken: [] }]
      parameters:
        - {
            name: role,
            in: query,
            required: true,
            schema: { $ref: "#/components/schemas/Role" },
            example: chat.fast,
          }
      responses:
        "200":
          description: Resolved model including the decrypted API key (service-to-service only).
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/SuccessResponse" }
                  - {
                      type: object,
                      properties:
                        {
                          data: { $ref: "#/components/schemas/ResolvedModel" },
                        },
                    }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/Error" }
        "400": { $ref: "#/components/responses/Error" }
        "503": { $ref: "#/components/responses/Error" }
components:
  securitySchemes:
    sessionCookie: { type: apiKey, in: cookie, name: thomas_session }
    csrfHeader: { type: apiKey, in: header, name: X-CSRF-Token }
    serviceToken: { type: http, scheme: bearer }
  responses:
    HealthStatus:
      description: Health result.
      content:
        application/json:
          schema:
            allOf:
              - { $ref: "#/components/schemas/SuccessResponse" }
              - { type: object, properties: { data: { type: object, required: [status], properties: { status: { type: string, enum: [ok, ready] } } } } }
    DeleteResult:
      description: Deletion confirmation.
      content:
        application/json:
          schema:
            allOf:
              - { $ref: "#/components/schemas/SuccessResponse" }
              - { type: object, properties: { data: { type: object, required: [deleted], properties: { deleted: { type: boolean } } } } }
    LogoutResult:
      description: Logout confirmation.
      content:
        application/json:
          schema:
            allOf:
              - { $ref: "#/components/schemas/SuccessResponse" }
              - { type: object, properties: { data: { type: object, required: [loggedOut], properties: { loggedOut: { type: boolean } } } } }
    ProviderTestResult:
      description: Provider connection test result.
      content:
        application/json:
          schema:
            allOf:
              - { $ref: "#/components/schemas/SuccessResponse" }
              - { type: object, properties: { data: { type: object, required: [ok, modelCount], properties: { ok: { type: boolean }, modelCount: { type: integer, minimum: 0 } } } } }
    RoleAssignment:
      description: Role assignment.
      content:
        application/json:
          schema:
            allOf:
              - { $ref: "#/components/schemas/SuccessResponse" }
              - { type: object, properties: { data: { $ref: "#/components/schemas/RoleAssignment" } } }
    Error:
      description: Error response.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    UnsupportedMediaType:
      description: The JSON body requires application/json; missing, malformed or other media types return UNSUPPORTED_MEDIA_TYPE.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Unauthorized:
      description: Authentication failed or is required.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Forbidden:
      description: The authenticated request is not allowed.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Provider:
      description: One provider.
      content:
        application/json:
          schema:
            allOf:
              - { $ref: "#/components/schemas/SuccessResponse" }
              - {
                  type: object,
                  properties:
                    { data: { $ref: "#/components/schemas/Provider" } },
                }
    ProviderList:
      description: Providers.
      content:
        application/json:
          schema:
            allOf:
              - { $ref: "#/components/schemas/SuccessResponse" }
              - {
                  type: object,
                  properties:
                    {
                      data:
                        {
                          type: array,
                          items: { $ref: "#/components/schemas/Provider" },
                        },
                    },
                }
    Model:
      description: One model.
      content:
        application/json:
          schema:
            allOf:
              - { $ref: "#/components/schemas/SuccessResponse" }
              - {
                  type: object,
                  properties: { data: { $ref: "#/components/schemas/Model" } },
                }
    ModelList:
      description: Models.
      content:
        application/json:
          schema:
            allOf:
              - { $ref: "#/components/schemas/SuccessResponse" }
              - {
                  type: object,
                  properties:
                    {
                      data:
                        {
                          type: array,
                          items: { $ref: "#/components/schemas/Model" },
                        },
                    },
                }
  schemas:
    Meta:
      type: object
      required: [requestId]
      properties: { requestId: { type: string } }
    SuccessResponse:
      type: object
      required: [data, meta]
      properties:
        data: {}
        meta: { $ref: "#/components/schemas/Meta" }
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              {
                type: string,
                description: "UPPER_SNAKE code from contracts/errors.yaml",
              }
            message: { type: string }
            details: { type: object }
            traceId: { type: string }
    User:
      type: object
      properties:
        {
          id: { type: string, format: uuid },
          email: { type: string },
          displayName: { type: string },
        }
    LoginRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        {
          email: { type: string, format: email },
          password: { type: string, format: password },
        }
    LoginSuccessResponse:
      allOf:
        - { $ref: "#/components/schemas/SuccessResponse" }
        - type: object
          properties:
            data:
              type: object
              properties:
                user: { $ref: "#/components/schemas/User" }
                csrfToken: { type: string }
                expiresAt: { type: string, format: date-time }
    ProviderKind:
      type: string
      enum:
        [
          openrouter,
          openai,
          anthropic,
          google,
          azure,
          ollama,
          openai_compatible,
          cohere,
          voyage,
          jina,
          typesafe,
        ]
    Capability:
      type: string
      enum: [chat, tools, vision, reasoning, embedding, rerank, decision]
    Role:
      type: string
      enum: [chat.default, chat.fast, vision, embedding, rerank, decision]
    Provider:
      type: object
      properties:
        id: { type: string, format: uuid }
        kind: { $ref: "#/components/schemas/ProviderKind" }
        name: { type: string }
        baseUrl: { type: string }
        hasApiKey: { type: boolean }
        enabled: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ProviderInput:
      type: object
      additionalProperties: false
      properties:
        kind: { $ref: "#/components/schemas/ProviderKind" }
        name: { type: string }
        baseUrl: { type: string }
        apiKey:
          {
            type: [string, "null"],
            writeOnly: true,
            description: "null/absent = unchanged unless baseUrl changes (then cleared), empty string = remove",
          }
        enabled: { type: boolean }
    Model:
      type: object
      properties:
        id: { type: string, format: uuid }
        providerId: { type: string, format: uuid }
        modelRef: { type: string }
        displayName: { type: string }
        capabilities:
          { type: array, items: { $ref: "#/components/schemas/Capability" } }
        contextWindow: { type: integer }
        embeddingDims: { type: integer }
        inputPricePerMTok: { type: number }
        outputPricePerMTok: { type: number }
        source: { type: string, enum: [manual, sync] }
    ModelInput:
      type: object
      additionalProperties: false
      required: [providerId, modelRef]
      properties:
        providerId: { type: string, format: uuid }
        modelRef: { type: string }
        displayName: { type: string }
        capabilities:
          { type: array, items: { $ref: "#/components/schemas/Capability" } }
        contextWindow: { type: integer }
        embeddingDims: { type: integer }
        inputPricePerMTok: { type: number }
        outputPricePerMTok: { type: number }
    RoleAssignment:
      type: object
      properties:
        {
          role: { $ref: "#/components/schemas/Role" },
          modelId: { type: string, format: uuid },
        }
    SyncResult:
      type: object
      properties: { added: { type: integer }, updated: { type: integer }, skipped: { type: integer } }
    ResolvedModel:
      type: object
      properties:
        role: { $ref: "#/components/schemas/Role" }
        fallbackFrom: { $ref: "#/components/schemas/Role" }
        providerKind: { $ref: "#/components/schemas/ProviderKind" }
        baseUrl: { type: string }
        apiKey: { type: string }
        modelRef: { type: string }
        capabilities:
          { type: array, items: { $ref: "#/components/schemas/Capability" } }
        embeddingDims: { type: integer }